api/app/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
confusable-homoglyphs==3.2.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-registration==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-nested-routers==0.92.5
entrypoints==0.3
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==2.0.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
